pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "pyarazzo"
description = "CLI to manipulate and build on top of the Arazzo Specification"
authors = [{name = "Sebastien Andreo", email = "sebastien@b-lab.io"}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.12"
keywords = []
dynamic = ["version"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1.8",
    "duty>=1.4.3",
    "httpx>=0.28.1",
    "jsonref>=1.1.0",
    "jsonschema>=4.23.0",
    "openapi-pydantic>=0.5.1",
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "robotframework>=7.4.0",
    "types-requests>=2.32.0.20250328",
]

[project.urls]
Homepage = "https://b-lab-io.github.io/pyarazzo"
Documentation = "https://b-lab-io.github.io/pyarazzo"
Changelog = "https://b-lab-io.github.io/pyarazzo/changelog"
Repository = "https://github.com/b-lab-io/pyarazzo"
Issues = "https://github.com/b-lab-io/pyarazzo/issues"
Discussions = "https://github.com/b-lab-io/pyarazzo/discussions"
Gitter = "https://gitter.im/b-lab-io/pyarazzo"
Funding = "https://github.com/sponsors/sebastien-b-lab-io"

[project.scripts]
pyarazzo = "pyarazzo.cli:main"

[tool.pdm]
version = {source = "scm"}

[tool.pdm.build]
package-dir = "src"
editable-backend = "editables"

# Include as much as possible in the source distribution, to help redistributors.
excludes = ["**/.pytest_cache"]
source-includes = [
    "config",
    "docs",
    "scripts",
    "share",
    "tests",
    "duties.py",
    "mkdocs.yml",
    "*.md",
    "LICENSE",
]

[tool.pdm.build.wheel-data]
# Manual pages can be included in the wheel.
# Depending on the installation tool, they will be accessible to users.
# pipx supports it, uv does not yet, see https://github.com/astral-sh/uv/issues/4731.
data = [
    {path = "share/**/*", relative-to = "."},
]

[dependency-groups]
dev = [
    # dev
    "editables>=0.5",
    "pytest",
    # maintenance
    "build>=1.2",
    "git-changelog>=2.5",
    "twine>=5.1",
    # ci
    "duty>=1.4",
    "ruff>=0.4",
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-randomly>=3.15",
    "pytest-xdist>=3.6",
    "mypy>=1.10",
    "types-markdown>=3.6",
    "types-pyyaml>=6.0",
    # docs
    "black>=24.4",
    "markdown-callouts>=0.4",
    "markdown-exec>=1.8",
    "mkdocs>=1.6",
    "mkdocs-coverage>=1.0",
    "mkdocs-gen-files>=0.5",
    "mkdocs-git-revision-date-localized-plugin>=1.2",
    "mkdocs-literate-nav>=0.6",
    "mkdocs-material>=9.5",
    "mkdocs-minify-plugin>=0.8",
    "mkdocstrings[python]>=0.25",
    # YORE: EOL 3.10: Remove line.
    "tomli>=2.0; python_version < '3.11'",
    "datamodel-code-generator>=0.26.4",
]

[tool.setuptools]
packages = ["pyarazzo"]

[tool.setuptools.package-data]
pyarazzo = ["*.yaml"]
